Initial refactor and linting
