+++ .swiftlint.yml
@@ -1,0 +1,2 @@
+disabled_rules:
+    - todo
