+++ PathKit.podspec
@@ -7,7 +7,7 @@ Pod::Spec.new do |spec|
   spec.author = { 'Kyle Fuller' => 'kyle@fuller.li' }
   spec.social_media_url = 'http://twitter.com/kylefuller'
   spec.source = { :git => 'https://github.com/kylef/PathKit.git', :tag => spec.version }
-  spec.source_files = 'Sources/PathKit.swift'
+  spec.source_files = 'Sources/PathKit*.swift'
   spec.ios.deployment_target = '8.0'
   spec.osx.deployment_target = '10.9'
   spec.tvos.deployment_target = '9.0'
